Moment Program

MomentPolynomialOpt.MOM.Model

MomentPolynomialOpt.MOM.MomentModel

MomentPolynomialOpt.MOM.constraint_moments

MomentPolynomialOpt.MOM.constraint_nneg

MomentPolynomialOpt.MOM.constraint_unitmass

MomentPolynomialOpt.MOM.constraint_zero

MomentPolynomialOpt.MOM.set_objective

MomentPolynomialOpt.MOM.set_objective_ncl

MomentPolynomialOpt.MOM.set_objective_tv

MomentPolynomialOpt.MOM.Model — Function

Construct the Moment Program in the variables X of order d. The moments of all monomials in X of degree 2*d are variables of the optimization program.

M = MOM.Model(X,d, optimizer=MMT[:optimizer]; nu=k)

X is the vector of variables

d is the order of the moment relaxation

'optimizer`is the the optimizer used to solve the SDP program

nu=k is the number of Positive Moment Sequences

M = MOM.Model( `sense`, f, [e1, e2, ...], [g1, g2, ...], X, d)

Construct the Moment Program in the variables X of order d.

sense == "inf" or "sup"

f polynomial objective function

[e1, e2, ...] array of polynomial equality constraints (can be empty)

[g1, g2, ...] array of non-negativity constraints (can be empty)

X is the vector of variables

d is the order of the moment relaxation.

M = MOM.Model(C, X, d)

Construct the Moment Program where

C is a vector of pairs (f, sense ) of objective or constraints where f is a polynomial and sense is "inf", "min", "sup", "max", ">=0", "<=0", "=0", or an interval

X is the vector of variables

d is the order of the moment relaxation.

MomentPolynomialOpt.MOM.MomentModel — Function

Construct the Moment Program in the variables X of order d. The moments of all monomials in X of degree 2*d are variables of the optimization program.

M = MomentModel(X,d; nu=k)

X is the vector of variables

d is the order of the moment relaxation.

nu=k is the number of Positive Moment Sequences

Constraints

MomentPolynomialOpt.MOM.constraint_zero — Function

constraint_zero(M, eqs ...)

Add to the moment program M, the constraints $e \star \mu_i = 0$ for i in 1:M[:nu] and e in eqs.

constraint_zero(M, i::Int, eqs ...)

Add to the moment program M, the constraint $e \star \mu_i = 0$ for e in eqs.

constraint_zero(M, Eqs ...)

Add to the moment program M, the constraints $\sum_{i=1}^{\nu} \mathbf{e}_{i} \star \mu_i =0$ for $\mathbf{e}$ in Eqs where Eqs is a sequence of vectors of $\nu$=M[:nu] polynomials.

MomentPolynomialOpt.MOM.constraint_nneg — Function

constraint_nneg(M, eqs ...)

Add to the moment program M, the constraints $e \star \mu_i \succeq 0$ for i in 1:M[:nu] and e in eqs.

constraint_nneg(M, i::Int64, eqs ...)

Add to the moment program M, the constraints $e \star \mu_i \succeq 0$ for e in eqs.

constraint_nneg(M, Eqs ...)

Add to the moment program M, the constraints $\sum_{i=1}^{\nu} \mathbf{e}_{i} \star \mu_i \succeq 0$ for $\mathbf{e}$ in Eqs where Eqs is a sequence of vectors of $\nu$=M[:nu] polynomials.

MomentPolynomialOpt.MOM.constraint_moments — Function

constraint_moments(M, [m => c, ...])

Add to the moment program M, the constraints $\langle \mu_i, m \rangle - c = 0$ for i in 1:M[:nu] and all pairs m=>c.

constraint_moments(M, [m => c, ...], p)

Add to the moment program M, the constraints $\sum_{i=1}^{\nu}\langle {p}_{i} \star \mu_i, m\rangle - c = 0$ for all pairs m=>c, where p is a vector of $\nu$=M[:nu] polynomials.

MomentPolynomialOpt.MOM.constraint_unitmass — Function

constraint_unitmass(M, i::Int64, p)

Add to the moment program M, the constraints $\langle {p} \star \mu_i, 1 \rangle - 1 = 0$ where p is a polynomial.

constraint_unitmass(M, p)

Add to the moment program M, the constraints $\langle {p}_i \star \mu_i, 1 \rangle - 1 = 0$ where p is a vector of $\nu$=M[:nu] polynomials.

Objective function

MomentPolynomialOpt.MOM.set_objective — Function

set_objective(M, sense, p)

Set the "inf" or "sup" objective function to $\sum_{i=1}^{\nu} \langle p\star \mu_i, 1 \rangle$ where p is a polynomial.

set_objective(M, sense, p)

Set the "inf" or "sup" objective function to $\sum_{i=1}^{\nu} \langle p_{i} \star \mu_i, 1 \rangle$ where p is a vector of $\nu$=M[:nu] polynomials.

set_objective(M, sense, i, p)

Set the "inf" or "sup" objective function to $\langle p \star \mu_i, 1 \rangle$ where p is a polynomial.

MomentPolynomialOpt.MOM.set_objective_tv — Function

set_objective_ncl(M)

Set the objective function of moment program to the total variation of the moment sequences $\mu_i$, that is the sum of the unit mass of the positive moment sequences $\sum_{i=1}^{\nu} \langle \mu_i, 1\rangle$.

MomentPolynomialOpt.MOM.set_objective_ncl — Function

set_objective_ncl(M)

Set the objective function of moment program to the nuclear norm or equivalently the sum of the traces of the moment matrices.