Moment Program

MomentPolynomialOpt.MOM.Model

MomentPolynomialOpt.MOM.add_constraint_moment

MomentPolynomialOpt.MOM.add_constraint_nneg

MomentPolynomialOpt.MOM.add_constraint_unitmass

MomentPolynomialOpt.MOM.add_constraint_zero

MomentPolynomialOpt.MOM.set_objective

MomentPolynomialOpt.MOM.set_objective_ncl

MomentPolynomialOpt.MOM.set_objective_tv

MomentPolynomialOpt.moments

MomentPolynomialOpt.moments — Function

moments(M, X, d::Int, symb::Symbol...)

Define variables for the moments of the monomials of degree less than d in the variables X. If

symb == :PSD, non-negativity constraints are added;

symb == :PRB, unit mass and non-negativity constraints are added.

MomentPolynomialOpt.MOM.Model — Function

M = MOM.Model( `sense`, f, H, G, X, d)

Construct the Moment Program in the variables X of order d.

sense == :inf or :sup

f polynomial objective function

H =[h1, h2, ...] array of polynomial equality constraints (can be empty)

G =[g1, g2, ...] array of non-negativity constraints (can be empty)

X is the vector of variables

d is the order of the moment relaxation.

M = MOM.Model(C, X, d)

Construct the Moment Program where

C is a vector of pairs (f, sense ) of objective or constraints where f is a polynomial and sense is "inf", "min", "sup", "max", ">=0", "<=0", "=0", or an interval

X is the vector of variables

d is the order of the moment relaxation.

Constraints

MomentPolynomialOpt.MOM.add_constraint_zero — Function

add_constraint_zero(M, eq, mu::Moments)

Add to the moment program M, the constraints $eq\star \mu = 0$, that is all the constraints $\mu(p*m)==0$ for $degree(m) \le 2*degree(M)-degree(p)$.

MomentPolynomialOpt.MOM.add_constraint_nneg — Function

add_constraint_nneg(M, mu::Moments, g)

Add to the moment program M, the constraints $g \star \mu \succeq 0$, that is the moment matrix of $g\star \mu$ is PSD.

MomentPolynomialOpt.MOM.add_constraint_moment — Function

add_constraint_moment(M, mu::Moments,  p => v)

Add to the moment program M, the constraint $\mu(p) - v == 0$.

MomentPolynomialOpt.MOM.add_constraint_unitmass — Function

add_constraint_unitmass(M, mu)

Add to the moment program M, the constraint $\langle \mu, 1 \rangle - 1 = 0$.

Objective function

MomentPolynomialOpt.MOM.set_objective — Function

set_objective(M, sense, p, mu)

Set the "inf" or "sup" objective function to $\sum_{i=1}^{\nu} \langle p\star \mu_i, 1 \rangle$ where p is a polynomial.

MomentPolynomialOpt.MOM.set_objective_tv — Function

set_objective_tv(M)

Set the objective function of the moment program to the total variation of the moment sequence $\mu$, that is the sum of the unit mass of the positive moment sequences $\sum_{i=1}^{\nu} \langle \mu_i, 1\rangle$.

MomentPolynomialOpt.MOM.set_objective_ncl — Function

set_objective_ncl(M)

Set the objective function of moment program to the nuclear norm or equivalently the trace of the moment matrices.